+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## v26.0.3 - /master current
 Release Date: UNRELEASED
-* Added `placeholder.com` image service. 
+* Issue 210: Added `Randomizer.Utf16String` that generates technically valid Unicode with paired high/low surrogates.
+* Added `placeholder.com` image service.
 
 ## v26.0.2
 Release Date: 2019-03-22